implement suggestions by @hsbadr
See
https://github.com/epiforecasts/EpiNow2/pull/213#discussion_r550701239
https://github.com/epiforecasts/EpiNow2/pull/213#discussion_r550701313
https://github.com/epiforecasts/EpiNow2/pull/213#discussion_r550701732

For now not implementing comparison to the approximate growth rate as
this seems quite a specific use case that could also be done outside the
stan model.

Also not implementing any approximate growth rate from seeding time -
instead minimum seeding time is now set to 1, so the last seeding time
is used to calculate the first growth rate.